feat(file-preview): Refactor to universal content viewer
Consolidate and enhance file content preview logic.

- Universal Viewer: Moved file content fetching and type handling entirely to `FileContentModal.vue`, making it a true universal viewer.
- Improved Type Detection: Implement content type determination that prioritizes MIME type (from fetched data) for display accuracy, falling back to extension for initial request routing.
- Centralized Definitions: Introduced `SUPPORTED_PREVIEW_TYPES` and refined `FileExtensionGroups`, `MimeGroups`, and `FileTypeIcons` for better organization and consistency.
- Eliminated Duplication: Removed redundant file content fetching logic from `CaseDashboard.vue`, resolving erroneous `400 Bad Request` calls for unsupported types.
- Enhanced Extensibility: The system is now more robust against misnamed files and easier to extend with new preview types.
